Incorporate PR feedback for rich text editor
Jira ticket: CAMS-526
