+++ user-interface/src/lib/components/cams/PrerenderedHtml/PrerenderedHtml.tsx
@@ -1,24 +1,5 @@
+import { DOMPURIFY_CONFIG } from '@/lib/utils/sanitize-text';
 import DOMPurify from 'dompurify';
-
-export const DOMPURIFY_CONFIG = {
-  ALLOWED_TAGS: ['em', 'strong', 'p', 'ul', 'ol', 'li', 'br', 'span', 'a'],
-  ALLOWED_ATTR: ['href', 'class', 'target', 'rel'],
-  ALLOW_DATA_ATTR: false,
-  FORBID_TAGS: ['style', 'script', 'iframe', 'object', 'embed', 'form', 'input', 'button'],
-  FORBID_ATTR: [
-    'onerror',
-    'onload',
-    'onclick',
-    'onmouseover',
-    'onmousedown',
-    'onfocus',
-    'onblur',
-    'onkeydown',
-    'onkeypress',
-    'srcset',
-    'src',
-  ],
-};
 
 type PrerenderedHtmlProps = {
   htmlString: string;
