+++ user-interface/src/lib/components/cams/RichTextEditor/RichTextEditor.scss
@@ -10,7 +10,6 @@
       gap: 0.25rem;
       padding: 0.5rem;
       border-bottom: 1px solid #e0e0e0;
-      //background-color: #f8f9fa;
       background-color: #ffffff;
       flex-wrap: wrap;
 
@@ -121,7 +120,6 @@
         overflow-y: auto;
       }
 
-      // Basic content styling
       p {
         margin: 0 0 0.5rem 0;
