+++ user-interface/src/lib/utils/sanitize-text.ts
@@ -4,13 +4,33 @@ import { IEventTelemetry } from '@microsoft/applicationinsights-web';
 import LocalStorage from './local-storage';
 import { filterToExtendedAscii } from '@common/cams/sanitization';
 
-const defaultConfiguration = {
+export const DOMPURIFY_CONFIG = {
+  ALLOWED_TAGS: ['em', 'strong', 'p', 'ul', 'ol', 'li', 'br', 'span', 'a'],
+  ALLOWED_ATTR: ['href', 'class', 'target', 'rel'],
+  ALLOW_DATA_ATTR: false,
+  FORBID_TAGS: ['style', 'script', 'iframe', 'object', 'embed', 'form', 'input', 'button'],
+  FORBID_ATTR: [
+    'onerror',
+    'onload',
+    'onclick',
+    'onmouseover',
+    'onmousedown',
+    'onfocus',
+    'onblur',
+    'onkeydown',
+    'onkeypress',
+    'srcset',
+    'src',
+  ],
+};
+
+const DOMPURIFY_TEXT_ONLY_CONFIG = {
   ALLOWED_TAGS: ['#text'],
   KEEP_CONTENT: false,
 };
 
 export function sanitizeText(maybeDirty: string, configuration: Partial<Config> = {}) {
-  const domPurifyConfig: Config = { ...defaultConfiguration, ...configuration };
+  const domPurifyConfig: Config = { ...DOMPURIFY_TEXT_ONLY_CONFIG, ...configuration };
   const clean = DOMPurify.sanitize(filterToExtendedAscii(maybeDirty), domPurifyConfig);
 
   if (DOMPurify.removed.length) {
